# Fix: Correctly load v8-compile-cache